- Removed 'find' from Factory -- we need a more versatile way to search for objects (predicates?) - Explicitly typed two variables in camera to suppress a warning :| - Misc.

+++ src/Camera.cpp
@@ -44,7 +44,7 @@ bool Camera::OnScreen(GameObject *object)
 {
    Square square(vector2(_position.x - (m_nScreenWidth * 0.5f), 
                          _position.y - (m_nScreenHeight * 0.5f)),
-                         m_nScreenWidth, m_nScreenHeight);
+                         (float)m_nScreenWidth, (float)m_nScreenHeight);
 
    ObjectState *objectState = object->getState();
